% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_treatment.R
\name{cap_at}
\alias{cap_at}
\title{Cap a numeric vector at a specified value}
\usage{
cap_at(x, cap, replacement = NA)
}
\arguments{
\item{x}{Numeric vector}

\item{cap}{Value above which to replace}

\item{replacement}{Value with which to replace elements above cap}
}
\description{
Replace all elements of a numeric vector that are above a specified value
with another value
}
