% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_treatment.R
\name{floor_at}
\alias{floor_at}
\title{Flor a numeric vector at a specified value}
\usage{
floor_at(x, floor, replacement = NA)
}
\arguments{
\item{x}{Numeric vector}

\item{floor}{Value below which to replace}

\item{replacement}{Value with which to replace elements below floor}
}
\description{
Replace all elements of a numeric vector that are below a specified value
with another value
}
