% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{normalize}
\alias{normalize}
\alias{normalise}
\title{Normalize a numeric vector}
\usage{
normalize(x)

normalise(...)
}
\arguments{
\item{x}{Numeric vector}
}
\description{
Special case of \code{\link{rescale}}() with new minimum of 0 and maximum of
1
}
