% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_treatment.R
\name{num_censor}
\alias{num_censor}
\title{Censor numeric vector}
\usage{
num_censor(x, bounds = c(-Inf, Inf))
}
\arguments{
\item{x}{Numeric vector}

\item{bounds}{Two-element numeric vector defining the lower and upper bounds
(in that order)}
}
\description{
Censor top and/or bottom of a numeric vector by replacing with the boundary
value(s)
}
\examples{
num_censor(0:10, c(2, 8))
}
