% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions.R
\name{num_iqr}
\alias{num_iqr}
\title{Describe numerc vector interquartile range (IQR)}
\usage{
num_iqr(x)
}
\arguments{
\item{x}{Numeric vector}
}
\description{
Calculate the interquartile range (IQR) of a numeric vector.
}
\examples{
num_iqr(1:10)
num_iqr(rnorm(1000, sd = 10))
}
