% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_transformations.R
\name{num_norm}
\alias{num_norm}
\title{Normalize a numeric vector}
\usage{
num_norm(x)
}
\arguments{
\item{x}{Numeric vector}
}
\description{
Special case of \code{\link{num_rescale}}() with new minimum of 0 and maximum
of 1
}
