% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{num_quantile}
\alias{num_quantile}
\title{Locate quantile boundaries of a numeric vector}
\usage{
num_quantile(x, probs = c(0, 1), na.rm = TRUE, names = FALSE, ...)
}
\arguments{
\item{x}{numeric vector whose sample quantiles are wanted, or an
    object of a class for which a method has been defined (see also
    \sQuote{details}). \code{\link{NA}} and \code{NaN} values are not
    allowed in numeric vectors unless \code{na.rm} is \code{TRUE}.}

\item{probs}{numeric vector of probabilities with values in
    \eqn{[0,1]}.  (Values up to \samp{2e-14} outside that
    range are accepted and moved to the nearby endpoint.)}

\item{na.rm}{logical; if true, any \code{\link{NA}} and \code{NaN}'s
    are removed from \code{x} before the quantiles are computed.}

\item{names}{logical; if true, the result has a \code{\link{names}}
    attribute.  Set to \code{FALSE} for speedup with many \code{probs}.}

\item{...}{further arguments passed to or from other methods.}
}
