% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{num_std}
\alias{num_std}
\title{Standardize a numeric vector}
\usage{
num_std(x, new_mean = 0, new_sd = 1)
}
\arguments{
\item{x}{Numeric vector}

\item{new_mean}{Numeric value of new mean}

\item{new_sd}{Numeric value of new standard deviation}
}
\description{
Standardize a numeric vector
}
