% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_treatment.R
\name{num_trunc}
\alias{num_trunc}
\title{Truncate numeric vector at specified value(s)}
\usage{
num_trunc(x, prop, tail = c("both", "top", "bottom"))
}
\arguments{
\item{x}{Numeric vector}

\item{prop}{Proportion of data to truncate}

\item{tail}{Tail on which to truncate}
}
\description{
Replace elements of a numeric vector identified by specific cut point(s)
}
\examples{
num_trunc(0:10, 2, 8)
}
