% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_treatment.R
\name{num_trunc_iqr}
\alias{num_trunc_iqr}
\title{Truncate numeric vector using interquartile range}
\usage{
num_trunc_iqr(x, iqr_multipler = 1.5)
}
\arguments{
\item{x}{Numeric vector}

\item{iqr_multipler}{Value by which to multiply the IQR when calculating the
target boundary}
}
\description{
Truncate elements of a numeric vector that fall outside the range defined by
the vector's mean, plus/minus the interquartile range (IQR) multipled by some
value (1.5 by default)
}
\examples{
boxplot(iris$Sepal.Width)  # Outliers present
boxplot(num_trunc_iqr(iris$Sepal.Width))  # Outliers removed

}
