% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_treatment.R
\name{num_truncate}
\alias{num_truncate}
\title{Truncate proportion of numeric vector}
\usage{
num_truncate(x, prop, tail = c("both", "top", "bottom"))
}
\arguments{
\item{x}{Numeric vector}

\item{prop}{Proportion of data to target}

\item{tail}{Tail(s) to target}
}
\description{
Truncate a proportion of elements from the top, bottom, or both tails of a
numeric vector by replacing them with NA
}
