% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_treatment.R
\name{p_to_quantiles}
\alias{p_to_quantiles}
\title{Convert single probability and tail method to two quantiles}
\usage{
p_to_quantiles(prop, tail = c("both", "top", "bottom"))
}
\description{
Convert single probability and tail method to two quantiles
}
