% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{rescale}
\alias{rescale}
\title{Rescale a numeric vector}
\usage{
rescale(x, new_min = 0, new_max = 1)
}
\arguments{
\item{x}{Numeric vector}

\item{new_min}{Numeric value of new minimum}

\item{new_max}{Numeric value of new maximum}
}
\description{
Rescale a numeric vector
}
