% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{standardise}
\alias{standardise}
\title{Standardise a numeric vector}
\usage{
standardise(x, new_mean = 0, new_sd = 1)
}
\arguments{
\item{x}{Numeric vector}

\item{new_mean}{Numeric value of new mean}

\item{new_sd}{Numeric value of new standard deviation}
}
\description{
Standardise a numeric vector
}
