% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{standardize}
\alias{standardize}
\alias{standardise}
\title{Standardize a numeric vector}
\usage{
standardize(x, new_mean = 0, new_sd = 1)

standardise(...)
}
\arguments{
\item{x}{Numeric vector}

\item{new_mean}{Numeric value of new mean}

\item{new_sd}{Numeric value of new standard deviation}
}
\description{
Standardize a numeric vector
}
