% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_treatment.R
\name{num_truncate}
\alias{num_truncate}
\alias{num_truncate_below}
\alias{num_truncate_above}
\title{Truncate numeric vector at specified value(s)}
\usage{
num_truncate(x, below = -Inf, above = Inf)

num_truncate_below(x, below)

num_truncate_above(x, above)
}
\arguments{
\item{x}{Numeric vector}

\item{below}{Value below which to truncate}

\item{above}{Value above which to truncate}
}
\description{
Replace elements of a numeric vector identified by specific cut point(s)
}
\examples{
num_truncate(0:10, 2, 8)
num_truncate_below(0:10, 2)
num_truncate_above(0:10, 8)

}
