% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_treatment.R
\name{num_truncate}
\alias{num_truncate}
\title{Truncate numeric vector}
\usage{
num_truncate(x, bounds = c(-Inf, Inf))
}
\arguments{
\item{x}{Numeric vector}

\item{bounds}{Two-element numeric vector defining the lower and upper bounds
(in that order)}
}
\description{
Truncate top and/or bottom of a numeric vector by replacing with NA
}
